Add Search, Inventories, fix friend bug (maybe)
Details on friend bug: Friends have to be serialized, but since Java
uses the same Reference User - Friendship < Requester (Orig User
reference) & Requested (Some other User reference), you get a cycle when
following Requester and trying to serialize his friendships. Solution is
to artificially limit depth of recursion by setting Requester to a deep
copy of the orig User but with friendships = null. This process was
broken slightly and was setting the requester and requested both to be
the original user. Is fixed in the simple case of getUserFromLogin, may
be broken elsewhere, something to keep an eye on.