Breakout Input Improvements
Added Keyboard input.

There are now 3 modes for input. First is following the pointer, second
is going left or right depending on the press location on the screen.
Third is keyboard.

Might need to remove one of the mouse modes, more testing needed on
mobile to confirm. Input might seem funny.